+++ src/ScottPlot/StarAxis.cs
@@ -1,17 +1,26 @@
 ﻿using ScottPlot.Drawing;
 using System;
-using System.Collections.Generic;
 using System.Drawing;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace ScottPlot
 {
     public struct StarAxisTick
     {
-        public double Location { get; set; }
-        public double[] Labels { get; set; }
+        public readonly double Location;
+        public readonly double[] Labels;
+
+        public StarAxisTick(double location, double[] labels)
+        {
+            Location = location;
+            Labels = labels;
+        }
+
+        public StarAxisTick(double location, double max)
+        {
+            Location = location;
+            Labels = new double[] { location * max };
+        }
     }
 
     public class StarAxis
@@ -33,9 +42,20 @@ namespace ScottPlot
         public string[] CategoryLabels;
 
         /// <summary>
+        /// Icons for each category.
+        /// Length must be equal to the number of columns (categories) in the original data.
+        /// </summary>
+        public Image[] CategoryImages;
+
+        /// <summary>
         /// Controls rendering style of the concentric circles (ticks) of the web
         /// </summary>
         public RadarAxis AxisType { get; set; }
+
+        /// <summary>
+        /// Indicates the type of axis chart to render
+        /// </summary>
+        public ImagePlacement ImagePlacement { get; set; }
 
         /// <summary>
         /// Color of the axis lines and concentric circles representing ticks
@@ -63,6 +83,7 @@ namespace ScottPlot
         /// Font used for labeling values on the plot
         /// </summary>
         public Drawing.Font Font = new();
+
         public int XAxisIndex { get; set; } = 0;
         public int YAxisIndex { get; set; } = 0;
 
@@ -70,21 +91,36 @@ namespace ScottPlot
         {
             double sweepAngle = 2 * Math.PI / NumberOfSpokes;
             double minScale = new double[] { dims.PxPerUnitX, dims.PxPerUnitX }.Min();
-            PointF origin = new PointF(dims.GetPixelX(0), dims.GetPixelY(0));
+            PointF origin = new(dims.GetPixelX(0), dims.GetPixelY(0));
 
-            using Pen pen = GDI.Pen(WebColor);
-            using StringFormat sf = new StringFormat() { LineAlignment = StringAlignment.Center };
-            using StringFormat sf2 = new StringFormat();
-            using System.Drawing.Font font = GDI.Font(Font);
-            using Brush fontBrush = GDI.Brush(Font.Color);
+            RenderRings(origin, minScale, sweepAngle);
+            RenderSpokes(origin, minScale, sweepAngle);
 
-            for (int i = 0; i < Ticks.Length; i++)
+            if (CategoryImages is not null)
             {
-                double tickDistancePx = Ticks[i].Location * minScale;
+                RenderImages(origin, minScale, sweepAngle);
+            }
+            else if (ShowCategoryLabels && CategoryLabels is not null)
+            {
+                RenderLabels(dims, origin, minScale, sweepAngle);
+            }
+        }
+
+        private void RenderRings(PointF origin, double minScale, double sweepAngle)
+        {
+            using Pen pen = GDI.Pen(WebColor);
+
+            foreach (var tick in Ticks)
+            {
+                double tickDistancePx = tick.Location * minScale;
 
                 if (AxisType == RadarAxis.Circle)
                 {
-                    Graphics.DrawEllipse(pen, (int)(origin.X - tickDistancePx), (int)(origin.Y - tickDistancePx), (int)(tickDistancePx * 2), (int)(tickDistancePx * 2));
+                    Graphics.DrawEllipse(pen,
+                        x: (int)(origin.X - tickDistancePx),
+                        y: (int)(origin.Y - tickDistancePx),
+                        width: (int)(tickDistancePx * 2),
+                        height: (int)(tickDistancePx * 2));
                 }
                 else if (AxisType == RadarAxis.Polygon)
                 {
@@ -98,10 +134,21 @@ namespace ScottPlot
                     Graphics.DrawPolygon(pen, points);
                 }
             }
+        }
+
+        private void RenderSpokes(PointF origin, double minScale, double sweepAngle)
+        {
+            using Pen pen = GDI.Pen(WebColor);
+            using System.Drawing.Font font = GDI.Font(Font);
+            using Brush fontBrush = GDI.Brush(Font.Color);
+            using StringFormat sf = new();
 
             for (int i = 0; i < NumberOfSpokes; i++)
             {
-                PointF destination = new PointF((float)(1.1 * Math.Cos(sweepAngle * i - Math.PI / 2) * minScale + origin.X), (float)(1.1 * Math.Sin(sweepAngle * i - Math.PI / 2) * minScale + origin.Y));
+                PointF destination = new(
+                    x: (float)(1.1 * Math.Cos(sweepAngle * i - Math.PI / 2) * minScale + origin.X),
+                    y: (float)(1.1 * Math.Sin(sweepAngle * i - Math.PI / 2) * minScale + origin.Y));
+
                 Graphics.DrawLine(pen, origin, destination);
 
                 for (int j = 0; j < Ticks.Length; j++)
@@ -115,36 +162,64 @@ namespace ScottPlot
                             float x = (float)(tickDistancePx * Math.Cos(sweepAngle * i - Math.PI / 2) + origin.X);
                             float y = (float)(tickDistancePx * Math.Sin(sweepAngle * i - Math.PI / 2) + origin.Y);
 
-                            sf2.Alignment = x < origin.X ? StringAlignment.Far : StringAlignment.Near;
-                            sf2.LineAlignment = y < origin.Y ? StringAlignment.Far : StringAlignment.Near;
+                            sf.Alignment = x < origin.X ? StringAlignment.Far : StringAlignment.Near;
+                            sf.LineAlignment = y < origin.Y ? StringAlignment.Far : StringAlignment.Near;
 
                             double val = Ticks[j].Labels[i];
-                            Graphics.DrawString($"{val:f1}", font, fontBrush, x, y, sf2);
+                            Graphics.DrawString($"{val:f1}", font, fontBrush, x, y, sf);
                         }
                         else if (i == 0)
                         {
                             double val = Ticks[j].Labels[0];
-                            Graphics.DrawString($"{val:f1}", font, fontBrush, origin.X, (float)(-tickDistancePx + origin.Y), sf2);
+                            Graphics.DrawString($"{val:f1}", font, fontBrush, origin.X, (float)(-tickDistancePx + origin.Y), sf);
                         }
                     }
                 }
             }
+        }
 
-            if (CategoryLabels != null && ShowCategoryLabels)
+        private void RenderImages(PointF origin, double minScale, double sweepAngle)
+        {
+            for (int i = 0; i < NumberOfSpokes; i++)
             {
-                for (int i = 0; i < NumberOfSpokes; i++)
-                {
+                double sweepOffset = ImagePlacement == ImagePlacement.Inside ? sweepAngle / 2 : 0;
+                double cosinus = Math.Cos(sweepAngle * i + sweepOffset - Math.PI / 2);
+                double sinus = Math.Sin(sweepAngle * i + sweepOffset - Math.PI / 2);
+                int imageWidth = CategoryImages[i].Width;
+                int imageHeight = CategoryImages[i].Height;
 
-                    PointF textDestination = new PointF(
-                        (float)(1.3 * Math.Cos(sweepAngle * i - Math.PI / 2) * minScale + origin.X),
-                        (float)(1.3 * Math.Sin(sweepAngle * i - Math.PI / 2) * minScale + origin.Y));
+                PointF imageDestination = new(
+                    (float)(1.45 * cosinus * minScale + origin.X - imageWidth / 2 * cosinus),
+                    (float)(1.45 * sinus * minScale + origin.Y - imageHeight / 2 * sinus));
 
-                    if (Math.Abs(textDestination.X - origin.X) < 0.1)
-                        sf.Alignment = StringAlignment.Center;
-                    else
-                        sf.Alignment = dims.GetCoordinateX(textDestination.X) < 0 ? StringAlignment.Far : StringAlignment.Near;
-                    Graphics.DrawString(CategoryLabels[i], font, fontBrush, textDestination, sf);
-                }
+                RectangleF rect = new(
+                    x: imageDestination.X - CategoryImages[i].Width / 2,
+                    y: imageDestination.Y - CategoryImages[i].Height / 2,
+                    width: CategoryImages[i].Width,
+                    height: CategoryImages[i].Height);
+
+                Graphics.DrawImage(CategoryImages[i], rect);
+            }
+        }
+
+        private void RenderLabels(PlotDimensions dims, PointF origin, double minScale, double sweepAngle)
+        {
+            using System.Drawing.Font font = GDI.Font(Font);
+            using Brush fontBrush = GDI.Brush(Font.Color);
+            using StringFormat sf = GDI.StringFormat(HorizontalAlignment.Center, VerticalAlignment.Middle);
+
+            for (int i = 0; i < NumberOfSpokes; i++)
+            {
+                PointF textDestination = new(
+                    (float)(1.3 * Math.Cos(sweepAngle * i - Math.PI / 2) * minScale + origin.X),
+                    (float)(1.3 * Math.Sin(sweepAngle * i - Math.PI / 2) * minScale + origin.Y));
+
+                if (Math.Abs(textDestination.X - origin.X) < 0.1)
+                    sf.Alignment = StringAlignment.Center;
+                else
+                    sf.Alignment = dims.GetCoordinateX(textDestination.X) < 0 ? StringAlignment.Far : StringAlignment.Near;
+
+                Graphics.DrawString(CategoryLabels[i], font, fontBrush, textDestination, sf);
             }
         }
     }